fix: hooks implementation
